+++ CHANGES.rst
@@ -278,9 +278,11 @@ astropy.units
 astropy.utils
 ^^^^^^^^^^^^^
 
-- ``astropy.utils.data.download_file`` now provides an ``http_headers`` keyword to
-  pass in specific request headers for the download. It also now defaults to
-  providing ``User-Agent: Astropy`` and ``Accept: */*`` headers.  [#9508]
+- ``astropy.utils.data.download_file`` and
+  ``astropy.utils.data.get_readable_fileobj`` now provides an ``http_headers``
+  keyword to pass in specific request headers for the download. It also now
+  defaults to providing ``User-Agent: Astropy`` and ``Accept: */*``
+  headers. [#9508]
 
 - Added a new ``astropy.utils.misc.unbroadcast`` function which can be used
   to return the smallest array that can be broadcasted back to the initial